requirements.txt
confluent-kafka
python-dotenv
redis
snowflake-connector-python
cryptography
